Aliases for common 'webapps'.
Taking advantage of the chromium 'app' feature. This launches the specified
webpage in full-screen mode (without forward, back or address bar) and makes
them feel like a web app on a phone or tablet.

Since this utility makes use of 'launch', the newly spawned processes exist
outside of the current process tree and belong to Init.