Utility script to record the runtime of an arbitrary command.
Although both time (built-in) and gnu time exist, it is not easy to isolate the
time output to store in a variable. This script suppresses all output
(including the trailing newline) except for the run time in seconds.

Useful when performing benchmarks.

TODO: cannot handle commands which spawn a subshell.