Improved reliability of runtime.
The uniqueness of the temporary file being used to store the execution time of a
command was being guaranteed by naming it using seconds since epoch. This could
cause difficulties should more than one user call the utility at the same
second.

The uniqueness was improved by using a name which includes both seconds since
epoch and current pid. This ensures that even if two users, (or the same user in
two different shells) calls the utility at the same time, it will still function
as expected.