Configuration files for using mutt for UNB email through davmail.
UNB email uses an exchange server. In order for mutt to work in this
environment, I'm using davmail locally to act as an adapter to the exchange
server, allowing me to access it as it it were a simple IMAP/SMTP mail server.

I'm using msmtp as my sendmail program.

Note that davmail must also be installed for this configuration to work
correctly. The setting:

	davmail.smtpStripFrom = true

was required in .davmail.properties to fix what seems to be an issue in davmail
itself (I read this on several forums).