terminal instances lauched through XMonad use client/daemon abilities of urxvt.
urxvt has the ability to run in daemon mode, where a single daemon is launched
on start-up, and all terminal instances are simply clients of the persistant
daemon. This configuration is said to use less RAM, but I'm interested in
the sharing of a single bash history between all instances.