I'm giving up on 80 char width text.
Trying to use constant 80 char width text is not practical for use with mobile devices and other settings. Also, the 80 char cut-off limit has been more of a nuisance than a helpful suggestion lately.

Lines are now not restricted to a certain width, and wrap on the nearest word boundary.